requirements.txt
Package             Version
------------------- --------
asgiref             3.6.0
autopep8            2.0.2
Django              4.1.7
django-filter       22.1
djangorestframework 3.14.0
Markdown            3.4.1
Pillow              9.4.0
pip                 22.3
pycodestyle         2.10.0
pytz                2022.7.1
setuptools          65.5.0
sqlparse            0.4.3
tzdata              2022.7
